refactor(GameActivity): GameActivity 관련 뷰 패키징 변경
