+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".GameActivity"
+            android:name=".gameactivity.GameActivity"
             android:exported="false" />
     </application>
