+++ app/src/main/java/woowacourse/omok/gameactivity/GameActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok
+package woowacourse.omok.gameactivity
 
 import android.content.Intent
 import android.os.Bundle
@@ -12,6 +12,8 @@ import domain.domain.OmokGame
 import domain.domain.Position
 import domain.domain.ProgressState
 import domain.library.combinerule.CombinedRuleAdapter
+import woowacourse.omok.FinishActivity
+import woowacourse.omok.R
 import woowacourse.omok.util.setOnSingleClickListener
 
 class GameActivity : AppCompatActivity() {
